requirements.txt
Flask==3.1.2
twilio==9.7.2
APScheduler==3.11.0
python-dotenv==1.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
